Update palword-rcon-monitored.bat
I changed the sequence of the Restart procedure because too much work when server shutting down. Remove the start server after shut down because the script check server running or not and it might be start before shut down process finished. Shutdown 0 not work for my server I change to "DoExit".